fix(billing): split settings owner controller
